feat(Tests): move model implementations to the Models namespace
